Move the nunit test to MSBuildDeviceIntegration test
